refactor: correct menu-open-close-icon path name
